refactor: extract canonical PDF hyperlinks
